requirements.txt
backoff
meraki>=1.0.0b6
PyYAML
redis
requests
websockets
sanic
Sanic-HTTPAuth
ssl